Debug

+++ test/utils/cleanup.ts
@@ -21,13 +21,13 @@ export default class CleanupUtils {
         const ps = [this.cmd(deleteCmd)]
         for (let img of data.images) {
             // https://lpt-aivideo-dst.s3.amazonaws.com/fgbFS5lIcd.gif
-            const path = img.url.split('https://')[1]
-            const bucketName = path.split('.s3.amazonaws.com/')[0]
-            const segs = path.split('/')
-            const objKey = segs[segs.length - 1]
-            console.log(`??? ${bucketName} ${objKey}`)
-
-            ps.push(this.cmd(`aws s3 rm s3://${bucketName}/${objKey}`))
+            if (img.url.startsWith(`https://lpt-aivideo-dst.s3.amazonaws.com`)) {
+                const path = img.url.split('https://')[1]
+                const bucketName = path.split('.s3.amazonaws.com/')[0]
+                const segs = path.split('/')
+                const objKey = segs[segs.length - 1]
+                ps.push(this.cmd(`aws s3 rm s3://${bucketName}/${objKey}`))
+            }
         }
         await Promise.all(ps)
     }
